substantially refactored components, need to add fixtures and test functionality
