+++ dashboard/app/routes/observability/episodes/EpisodesTable.tsx
@@ -1,5 +1,4 @@
 import { useState } from "react";
-import type { EpisodeByIdRow, TableBounds } from "~/utils/clickhouse";
 import {
   Table,
   TableBody,
@@ -10,20 +9,14 @@ import {
 } from "~/components/ui/table";
 import { Button } from "~/components/ui/button";
 import { Input } from "~/components/ui/input";
-import { ChevronLeft, ChevronRight } from "lucide-react";
-import { useNavigate } from "react-router";
+import type { EpisodeByIdRow } from "~/utils/clickhouse/inference";
 
 export default function EpisodesTable({
   episodes,
-  pageSize,
-  bounds,
 }: {
   episodes: EpisodeByIdRow[];
-  pageSize: number;
-  bounds: TableBounds;
 }) {
   const [goToId, setGoToId] = useState("");
-  const navigate = useNavigate();
 
   // TODO: wire this to go the the details page for a particular inference, maybe add a popover.
   const handleGoTo = (e: React.FormEvent<HTMLFormElement>) => {
@@ -49,24 +42,6 @@ export default function EpisodesTable({
     const end = endTime.toLocaleString("en-US", formatOptions);
     return `${start} — ${end}`;
   };
-
-  const topEpisode = episodes[0];
-  const bottomEpisode = episodes[episodes.length - 1];
-
-  // IMPORTANT: use the last_inference_id to navigate
-  const handleNextPage = () => {
-    navigate(
-      `?before=${bottomEpisode.last_inference_id}&page_size=${pageSize}`,
-    );
-  };
-
-  const handlePreviousPage = () => {
-    navigate(`?after=${topEpisode.last_inference_id}&page_size=${pageSize}`);
-  };
-
-  // These are swapped because the table is sorted in descending order
-  const disablePrevious = bounds.last_id === topEpisode.last_inference_id;
-  const disableNext = bounds.first_id === bottomEpisode.last_inference_id;
 
   return (
     <div>
@@ -97,7 +72,10 @@ export default function EpisodesTable({
           {episodes.map((episode) => (
             <TableRow key={episode.episode_id} id={episode.episode_id}>
               <TableCell className="max-w-[200px] lg:max-w-none">
-                <a href="#" className="block no-underline">
+                <a
+                  href={`/observability/episode/${episode.episode_id}`}
+                  className="block no-underline"
+                >
                   <code className="block overflow-hidden text-ellipsis whitespace-nowrap rounded font-mono transition-colors duration-300 hover:text-gray-500">
                     {episode.episode_id}
                   </code>
@@ -117,22 +95,6 @@ export default function EpisodesTable({
           ))}
         </TableBody>
       </Table>
-      <div className="mt-4 flex items-center justify-center gap-2">
-        <Button
-          onClick={handlePreviousPage}
-          disabled={disablePrevious}
-          className="rounded-md border border-gray-300 bg-white p-2 text-sm font-medium text-gray-700 hover:bg-gray-50"
-        >
-          <ChevronLeft className="h-4 w-4" />
-        </Button>
-        <Button
-          onClick={handleNextPage}
-          disabled={disableNext}
-          className="rounded-md border border-gray-300 bg-white p-2 text-sm font-medium text-gray-700 hover:bg-gray-50"
-        >
-          <ChevronRight className="h-4 w-4" />
-        </Button>
-      </div>
     </div>
   );
 }
