+++ dashboard/app/utils/clickhouse/common.test.ts
@@ -1,0 +1,7 @@
+import { expect, test } from "vitest";
+import { checkClickhouseConnection } from "./common";
+
+test("checkClickhouseConnection", async () => {
+  const result = await checkClickhouseConnection();
+  expect(result).toBe(true);
+});
