+++ dashboard/app/utils/date.ts
@@ -1,0 +1,9 @@
+export const formatDate = (date: Date) => {
+  return date.toLocaleString("en-US", {
+    year: "numeric",
+    month: "short",
+    day: "numeric",
+    hour: "2-digit",
+    minute: "2-digit",
+  });
+};
