+++ dashboard/app/utils/supervised_fine_tuning/fireworks.ts
@@ -23,18 +23,18 @@
 import type { SFTFormValues } from "~/routes/optimization/supervised-fine-tuning/types";
 import type { JsExposedEnv } from "../minijinja/pkg/minijinja_bindings";
 import { v7 } from "uuid";
-import type {
-  ContentBlockOutput,
-  JsonInferenceOutput,
-  ParsedInferenceRow,
-} from "../clickhouse";
-import { getCuratedInferences } from "../clickhouse";
 import { render_message } from "./rendering";
 import { getConfig } from "../config/index.server";
 import { get_template_env, type ChatCompletionConfig } from "../config/variant";
 import { z } from "zod";
 import { SFTJob, type SFTJobStatus } from "./common";
 import type { ProviderType } from "../config/models";
+import { getCuratedInferences } from "../clickhouse/curation";
+import type {
+  ContentBlockOutput,
+  JsonInferenceOutput,
+  ParsedInferenceRow,
+} from "../clickhouse/common";
 export const FIREWORKS_API_URL = "https://api.fireworks.ai";
 export const FIREWORKS_API_KEY = process.env.FIREWORKS_API_KEY || logError();
 export const FIREWORKS_ACCOUNT_ID =
